[refactor] 팔로워, 팔로잉 가져오는 서비스 이름 변경
