+++ src/main/java/com/hicc/nagne_backend/domain/follow/application/service/FollowerGetService.java
@@ -3,7 +3,7 @@ package com.hicc.nagne_backend.domain.follow.application.service;
 import com.hicc.nagne_backend.domain.follow.application.dto.FollowResponse;
 import com.hicc.nagne_backend.domain.follow.application.mapper.FollowMapper;
 import com.hicc.nagne_backend.domain.follow.domain.entity.Follow;
-import com.hicc.nagne_backend.domain.follow.domain.service.FollowingQueryService;
+import com.hicc.nagne_backend.domain.follow.domain.service.FollowerQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,12 +11,13 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class ReceiverGetService {
+public class FollowerGetService {
 
-    private final FollowingQueryService followingQueryService;
+    private final FollowerQueryService followerQueryService;
 
-    public List<FollowResponse.FollowInfoResponse> getReceiver(Long userId) {
-        List<Follow> receiverIdList = followingQueryService.findFollowingListById(userId);
-        return FollowMapper.mapToFollowingList(receiverIdList);
+    public List<FollowResponse.FollowInfoResponse> getSender(Long userId) {
+        List<Follow> receiverIdList = followerQueryService.findFollowerListById(userId);
+        return FollowMapper.mapToFollowerList(receiverIdList);
+
     }
 }
