+++ src/main/java/com/hicc/nagne_backend/domain/follow/presentation/FollowController.java
@@ -1,8 +1,8 @@
 package com.hicc.nagne_backend.domain.follow.presentation;
 
 import com.hicc.nagne_backend.domain.follow.application.dto.FollowResponse;
-import com.hicc.nagne_backend.domain.follow.application.service.ReceiverGetService;
-import com.hicc.nagne_backend.domain.follow.application.service.SenderGetService;
+import com.hicc.nagne_backend.domain.follow.application.service.FollowerGetService;
+import com.hicc.nagne_backend.domain.follow.application.service.FollwingGetService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,17 +14,17 @@ import java.util.List;
 @RequiredArgsConstructor
 public class FollowController {
 
-    private final SenderGetService senderGetService;
-    private final ReceiverGetService receiverGetService;
+    private final FollowerGetService followerGetService;
+    private final FollwingGetService followingGetService;
 
     @GetMapping("/followSender/{userId}")
     public List<FollowResponse.FollowInfoResponse> getSender(@PathVariable Long userId){
-        return senderGetService.getSender(userId);
+        return followerGetService.getSender(userId);
     }
 
     @GetMapping("/followReceiver/{userId}")
     public List<FollowResponse.FollowInfoResponse> getReceiver(@PathVariable Long userId){
-        return receiverGetService.getReceiver(userId);
+        return followingGetService.getReceiver(userId);
     }
 
 }
